[Feat]#18 - MainView 생성
