+++ PennyPack/Info.plist
@@ -2,6 +2,11 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
+	<key>APIKey
+currency_APIKey
+currency_APIKey
+currency_APIKey</key>
+	<string>$(APIKey)</string>
 	<key>CFBundleURLTypes</key>
 	<array>
 		<dict>
@@ -9,10 +14,11 @@
 			<string></string>
 		</dict>
 	</array>
-	<key>APIKey
-currency_APIKey
-currency_APIKey
-currency_APIKey</key>
-	<string>$(APIKey)</string>
+	<key>UIAppFonts</key>
+	<array>
+		<string>GmarketSansBold.otf</string>
+		<string>GmarketSansMedium.otf</string>
+		<string>SpoqaHanSansNeo-Bold</string>
+	</array>
 </dict>
 </plist>
