+++ PennyPack/ViewModel/OtherFuncs.swift
@@ -1,0 +1,18 @@
+import Foundation
+import SwiftUI
+
+// MARK: RoundedCorner custom
+struct RoundedCorner: Shape {
+    var radius: CGFloat = 0.0
+    var corners: UIRectCorner = .allCorners
+
+    func path(in rect: CGRect) -> Path {
+        let path = UIBezierPath(
+            roundedRect: rect,
+            byRoundingCorners: corners,
+            cornerRadii: CGSize(width: radius, height: radius)
+        )
+        return Path(path.cgPath)
+    }
+}
+
